requirements.txt
# ================================================================
# REQUIREMENTS.TXT - Dependencias del proyecto ETL Meteorológico
# ================================================================

# Manipulación de datos
pandas==2.1.4
numpy==1.26.2

# Conexión a PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Peticiones HTTP (API)
requests==2.31.0
urllib3==2.1.0

# Variables de entorno
python-dotenv==1.0.0

# Utilidades de fecha/hora
python-dateutil==2.8.2
pytz==2023.3

# Opcional: Para análisis de datos
matplotlib==3.8.2
seaborn==0.13.0

# Opcional: Para logging avanzado
colorlog==6.8.0

# ================================================================
# INSTALACIÓN:
# pip install -r requirements.txt
# ================================================================